+++ 114408-front-end/lib/types/HtmlDivType.ts
@@ -1,0 +1,1 @@
+export type HtmlDivPropsType = React.HTMLAttributes<HTMLDivElement>;
